chore: add commission payments type to client
