+++ npm-admin-client/types/markets.ts
@@ -130,9 +130,3 @@ export type MarketAccount = {
 };
 
 export type EpochTimeStamp = number;
-
-export type PaymentInfo = {
-  from: PublicKey;
-  to: PublicKey;
-  amount: BN;
-};
